Improved method readability and added comments
