+++ config/default/migrate_plus.migration.hs_courses.yml
@@ -16,6 +16,7 @@ migration_group: hs_courses
 label: 'Course importer'
 source:
   track_changes: true
+  delete_orphans: true
   plugin: url
   data_fetcher_plugin: http
   urls: 'http://events.stanford.edu/xml/drupal/v2.php?organization=0'
