+++ docroot/modules/humsci/hs_migrate/hs_migrate.services.yml
@@ -1,6 +1,6 @@
 services:
-  hs_migrate.migrate_events:
-    class: Drupal\hs_migrate\EventSubscriber\MigrateEventsSubscriber
+  hs_migrate.event_subscriber:
+    class: Drupal\hs_migrate\EventSubscriber\EventsSubscriber
     arguments: ['@entity_type.manager']
     tags:
       - { name: event_subscriber }
